Se actualiza vaper product  y se hacen las relaciones de integridad referencial de la base
Se actualiza vaper product  y se hacen las relaciones de integridad referencial de la base, se quitan cosas innecesarias
